Input validation testing of native GCMSIV.

+++ native_c/intel/packet/gcm_siv_pc/gcm_siv_pc.c
@@ -4,7 +4,7 @@
 #include <memory.h>
 
 packet_err *
-gcm_siv_pc_process_packet(bool encryption, uint8_t *key, size_t keysize, uint8_t *nonce, uint8_t *aad, int aadLen,
+gcm_siv_pc_process_packet(bool encryption, uint8_t *key, size_t keysize, uint8_t *nonce, uint8_t *aad, size_t aadLen,
                           uint8_t *p_in, size_t inLen, uint8_t *p_out, size_t *outputLen) {
     __m128i roundKeys[15];
     __m128i theGHash = _mm_setzero_si128();
@@ -23,7 +23,7 @@ gcm_siv_pc_process_packet(bool encryption, uint8_t *key, size_t keysize, uint8_t
     }
     gcm_siv_hasher_completeHash(&theAEADHasher, T, &theGHash);
     if (encryption) {
-        gcm_siv_hasher_updateHash(&theDataHasher, T, p_in, (int) inLen, &theGHash);
+        gcm_siv_hasher_updateHash(&theDataHasher, T, p_in, inLen, &theGHash);
         calculateTag(&theDataHasher, &theAEADHasher, T, roundKeys, &theGHash, (int8_t *) nonce, macBlock, &p_encrypt);
         gcm_siv_process_packet(p_in, (int) inLen, macBlock, roundKeys, p_out, &p_encrypt);
         memcpy(p_out + inLen, macBlock, BLOCK_SIZE);
@@ -31,7 +31,7 @@ gcm_siv_pc_process_packet(bool encryption, uint8_t *key, size_t keysize, uint8_t
     } else {
         *outputLen = inLen - BLOCK_SIZE;
         gcm_siv_process_packet(p_in, (int) *outputLen, p_in + *outputLen, roundKeys, p_out, &p_encrypt);
-        gcm_siv_hasher_updateHash(&theDataHasher, T, p_out, (int) *outputLen, &theGHash);
+        gcm_siv_hasher_updateHash(&theDataHasher, T, p_out,  *outputLen, &theGHash);
         calculateTag(&theDataHasher, &theAEADHasher, T, roundKeys, &theGHash, (int8_t *) nonce, macBlock, &p_encrypt);
         if (!tag_verification_16(macBlock, p_in + *outputLen)) {
             memset(p_out, 0, *outputLen);
